Save resource whenever constraints are saved in DeployView.
This improves the problem that changes in DeployView are not shown in Graphiti. Now, whenever the name or assigned place for a FunctionBlockModel get changed in DeployView and Graphiti is reloaded, the graphiti diagram already has the new values. Only problem is, it doesn't display them, but displays the old values in red.