Implement some autoconfiguration for Eclipse
In reference to bug #13

Eclipse now does autoconfiguration for network addresses. It defaults
to listening on all addresses on port 5000 and usese 225.0.0.1:5000 on
all devices as its multicast address. It will also automatically set
the announce addresses either based on the listen addresses or all
addresses that can be found.

It's not perfect and the code should be refactored as ModuleMain is
still missing autoconfig support and the code would have to be
duplicated otherwise