One may now enjoy enhanced sorting and killing functionality while using the splendidly designed AppView.
Minor flaws shall be reported and are not to be seen as signs of failure,
but be admired as the most pure excellence and genius of the programmer,
due to which she failed to foresee one's potential incompetence while
using this perfectly assembled view.
She will then show her gratitude for your advice and mercy upon your
stupidity, do her best to fix whatever problems and fears you might encounter,
and enlighten your ways through the dire and dangerous forests
that darken the graphical user interface.