AppView is able to display the names of running function blocks in addition to their types. I had to change a lot of classes, since this information has to be passed through several messages, wrappers and other stuff, but only had to add small parts. I didn't change or remove any existing code.
I noticed that whenever a function block model is deployed,
a corresponding block is created with the same UUID. I think
it would be better if every block had a unique UUID, even if
they derive from the same block model. Think of differing thresholds
after changing one block of an application, it would be nice
to be able to distinguish between blocks that have the same type
but different thresholds, since they represent different states
of the application.
The only advantage I see in having the same UUID both for the model
and the block is being able to connect them later on, but I can't
think of any task where we'd actually need to use that information.