Fixed bug that caused inconsistency between DeployView and Graphiti after - creating block in graphiti & saving - loading blocks into DeployView - deleting block in graphiti & saving - editing constraints for the deleted block, saving constraints.
Now whenever Constraints are submitted to be saved, the blocks get updated
first. If the block that should get the new constraints has been removed,
no constraints are saved, the user receives a warning message and everything
works. If the block still is there, the constraints are saved. In any case,
the block table is now updated whenever the save constraints button is pressed.