Fix bug regarding free block type holders
When an application was deployed twice (or just any two or more applications
wit blocks that have the same UUID) then after shutting down the first the
others wouldn't free the BlockTypeHolders they used. Also only the
BlockTypeHolder for the last block with a given UUID added would be freed.