IMPORTANT COMMIT CONCERNING BLOCK-SYNCHRO BETWEEN GRAPHITI AND DEPLOYVIEW
Removed the "isAutoUpdateAtStartup()" Method from DNDDiagramTypeProvider.
This got rid of the necessity to save or not save the diagram every single
time Graphiti was opened, even if the diagram hadn't been changed at all.

I don't know yet whether this will affect the Synchronization of the
BlockModels between Graphiti and DeployView. If it does, just insert the
method again, although it is a little annoying for the user to have to save
the diagram so many times without it actually having changed.